+++ src/searchad/reports/ReportingRequest.php
@@ -33,6 +33,8 @@ use searchad\ApiRequest;
  * adminArea
  * deviceClass
  * ageRange
+ * locality
+ * gender
  * @property array $groupBy
  *
  * @property bool $returnRowTotals Specify whether to return total of each row. Default is false.
@@ -254,16 +256,23 @@ class ReportingRequest extends ApiRequest
     }
 
     /**
-     * @param  string $groupField
+     * Geo params should also be set in condition query
+     * @param  string|array $groupField
      * @throws \Exception
      */
     public function setGroupBy($groupField)
     {
-        $possible = ['countryCode', 'adminArea', 'deviceClass', 'ageRange'];
-        if (!in_array($groupField, $possible)) {
-            throw  new \Exception("Group by field is invalid");
+        $possible = ['countryCode', 'adminArea', 'deviceClass', 'ageRange', 'locality', 'gender'];
+        $groupFields = is_string($groupField) ? [$groupField] : $groupField;
+        $validFields = [];
+        foreach($groupFields as $field){
+            if (!in_array($field, $possible)) {
+                throw  new \Exception("Group by field is invalid");
+            }
+            $validFields[] = $field;
         }
-        $this->groupBy = [$groupField];
+
+        $this->groupBy = $validFields;
     }
 
     /**